fix: build routed-expert range check and default route in int64 for uint16 ids
Models with more than 255 experts carry routed-expert ids as uint16, and torch
has no max()/arange() CPU kernels for the unsigned 16/32-bit dtypes. With router
replay enabled on a 512-expert model every sample in the async trajectory
collector raised NotImplementedError("max_all" not implemented for 'UInt16')
in the overflow guard of _as_routed_experts_tensor, and the default-route
padding in pad_and_align_routed_expert_indices raises the same for arange.

Reduce in int64 before comparing against iinfo(dtype).max, and build the
default route in int64 before narrowing to the carry dtype. The narrowing casts
are unchanged, so the guard still catches a mis-detected expert count.

Adds a unit test that feeds uint16 ids (max 511) with padding through
pad_and_align_routed_expert_indices.

+++ nemo_rl/models/generation/vllm/utils.py
@@ -140,7 +140,10 @@ def _as_routed_experts_tensor(
     global G_ROUTED_EXPERTS_RANGE_CHECKED
     tensor = torch.as_tensor(value, device=device)
     if not G_ROUTED_EXPERTS_RANGE_CHECKED and tensor.numel() > 0:
-        max_id = int(tensor.max())
+        # Reduce in int64: torch has no max() kernel for the unsigned carry
+        # dtypes (uint16/uint32) that models with more than 255 experts use, so
+        # ``tensor.max()`` raises NotImplementedError("max_all" ...) on them.
+        max_id = int(tensor.to(torch.int64).max())
         limit = torch.iinfo(dtype).max
         if max_id > limit:
             raise ValueError(
@@ -307,11 +310,13 @@ def pad_and_align_routed_expert_indices(
             "Router replay allows at most one surplus final-token route."
         )
 
+    # Build in int64 and narrow afterwards: torch has no arange kernel for the
+    # unsigned carry dtypes (uint16/uint32) used when a model has > 255 experts.
     default_route = torch.arange(
         routed.shape[2],
-        dtype=routed_experts_dtype,
+        dtype=torch.int64,
         device=device,
-    )
+    ).to(dtype=routed_experts_dtype)
     full = (
         default_route.view(1, 1, -1)
         .expand(padded_length, routed.shape[1], routed.shape[2])
